server/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
idna==2.8
python-decouple==3.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.3
